// src/api/event/eventModel.ts
export const CreateSessionEventSchema = z.object({
    level: z.number(),
    time: z.number(),
    pid: z.number(),
    hostname: z.string(),
    name: z.string(),
    type: z.string(),
    sourceChannel: z.string().optional(),
    sourceNumber: z.string().optional(),
    queue: z.string().optional(),
    destChannel: z.string().optional(),
    destNumber: z.string().optional(),
    date: z.string().refine(value => !isNaN(Date.parse(value)), {
        message: "Invalid date format",
    }),
    duration: z.string(),
    filename: z.string(),
    msg: z.string(),
});

const TranscriptionSegmentSchema = z.object({
    start: z.number(),
    end: z.number(),
    speaker: z.string(),
    text: z.string(),
});

export const TranscriptionResponseSchema = z.object({
    wav_customer: z.array(TranscriptionSegmentSchema),
});

export type TranscriptionResponse = z.infer<typeof TranscriptionResponseSchema>;

export const AnalysisResponseSchema = z.object({
    analysis: z.object({
        explanation: z.array(z.string()),
        category: z.array(z.string()),
        topic: z.record(z.string(), z.string()),
        emotion: z.array(z.string()),
        key_words: z.array(z.string()),
        routin_check_start: z.array(z.string()),
        routin_check_end: z.array(z.string()),
        forbidden_words: z.record(z.string(), z.number()), // Changed to record with number values
    })
});